Update load-balance-access-application-cluster.md
It seems there is a difference in handling MD format on github and kubernetes.io site. On k8s site "curl <your-external-ip-address>:8080" shows up as "curl :8080" as "<your-external-ip-address>" treated as html tag. It also resets section numbering (Use your Service object to access the Hello World application: is marked as "1." while it should be "5.")
"curl <minikube-node-ip-address>:<service-node-port>" seems to be treated correctly but I hope extra explicit formatting will not break it. Alas Preview on github is not very useful to validate it.